requirements.txt
aiogram==3.20.0.post0
aiosqlite>=0.19.0
sentence-transformers>=2.3.0
openai==0.28.1
python-dotenv==1.0.0
numpy>=1.21.0
huggingface_hub>=0.16.0,<0.20.0
qdrant-client>=1.7.0
loguru~=0.7.3